[WIP] Added context menu actions and updated tests
